requirements.txt
playwright==1.40.0
pydantic==2.5.3
python-dotenv==1.0.0
pyodbc==5.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
